refactor(main): extract command facades into cmd_facades (main.py god-file slice R5)
